+++ app/routes/components/Select/changelog.codegen.ts
@@ -12,4 +12,5 @@ export const changelog = ''
   + '- 2025-10-26: Update styles (padding) to make it look better.\n'
   + '- 2025-10-27: Update styles when Select is opened or focused. Make its\n'
   + '  `min-width` be `8rem` now (rather than `16rem`).\n'
+  + '- 2025-10-28: Refactor code to make code cleaner. Add a small gap in trigger.\n'
   ;