+++ src/components/Select/select.tsx
@@ -1,11 +1,10 @@
-import classNames from "classnames";
-import { ChevronsUpDown } from "lucide-react";
 import { useEffect, useLayoutEffect, useRef, useState } from "react";
 
 import { useJss } from "@/components/CopyUiProvider";
 import { Popover } from "@/components/Popover";
-import { useTheme } from "@/components/ThemeProvider";
-import styles from "./select.module.scss";
+
+import { SelectList } from "./select-list";
+import { SelectTrigger } from "./select-trigger";
 
 interface SelectProps<V extends string> {
   id?: string;
@@ -32,10 +31,9 @@ const Select = <V extends string>(props: SelectProps<V>) => {
     onChange,
   } = props;
 
-  const theme = useTheme();
   const jss = useJss();
 
-  const mainRef = useRef<HTMLButtonElement | null>(null);
+  const triggerRef = useRef<HTMLButtonElement | null>(null);
 
   const [internalValue, setInternalValue] = useState<V | null>(defaultValue);
   const [mainWidth, setMainWidth] = useState(0);
@@ -47,119 +45,61 @@ const Select = <V extends string>(props: SelectProps<V>) => {
 
   // Update main width on mount.
   useLayoutEffect(() => {
-    const main = mainRef.current;
+    const main = triggerRef.current;
     if (main) {
       setMainWidth(main.offsetWidth);
     }
   }, []);
 
   const showLabel = internalValue
-    ? options.find((o) => o.value === internalValue)?.label
+    ? (options.find((o) => o.value === internalValue)?.label ?? null)
     : null;
 
-  const baseStyle = {
+  const baseStx = jss.hash({
     "--select-list-padding": "0.25rem",
     "--select-list-item-padding-inline": "0.5rem",
     "--select-main-padding-inline":
       "calc(var(--select-list-padding) + var(--select-list-item-padding-inline))",
-  };
-
-  const mainStx = jss.hash({
-    ...baseStyle,
-
-    "--select-font-size": theme.tokens.inputBaseMdFontSize,
-    "--select-line-height": theme.tokens.inputBaseMdLineHeight,
-
-    "--select-main-height": theme.tokens.inputBaseMdHeight,
-    "--select-main-border-color": theme.tokens.inputBaseDefaultBorderColor,
-    "--select-main-border-radius": theme.tokens.inputBaseBorderRadius,
-    "--select-main-placeholder-color": theme.tokens.inputBasePlaceholderColor,
-    "--select-main-disabled-color": theme.colors.gray["600"],
-    "--select-main-disabled-bg-color": theme.colors.gray["000"],
-    "--select-main-disabled-border-color": theme.colors.gray["200"],
-    "--select-main-focus-border-color": theme.colors.blue["800"],
-  });
-
-  const listStx = jss.hash({
-    ...baseStyle,
-
-    "--select-font-size": theme.tokens.inputBaseMdFontSize,
-    "--select-line-height": theme.tokens.inputBaseMdLineHeight,
-
-    "--select-list-width": `${mainWidth}px`,
-    "--select-list-border-color": theme.tokens.inputBaseDefaultBorderColor,
-    "--select-list-border-radius": theme.tokens.inputBaseBorderRadius,
-
-    "--select-item-hover-bg-color": theme.colors.gray["000"],
   });
 
   return (
     <Popover>
       <Popover.Trigger
         render={({ setRef, isOpen, onToggle }) => (
-          <button
+          <SelectTrigger
             id={id}
             ref={(el) => {
               setRef(el);
-              mainRef.current = el;
+              triggerRef.current = el;
             }}
-            className={classNames(styles.selectMain, mainStx)}
-            onClick={() => !disabled && onToggle()}
-            data-value-picked={internalValue == null ? undefined : "true"}
-            data-disabled={disabled ? "true" : undefined}
-            data-opened={isOpen ? "true" : undefined}
-            type="button"
-            role="combobox"
-            aria-haspopup="listbox"
-            aria-expanded="false"
-          >
-            <span style={{ flex: 1 }}>
-              {showLabel ? showLabel : placeholder}
-            </span>
-            <span className={styles.selectIcon}>
-              <ChevronsUpDown size="62.5%" />
-            </span>
-          </button>
+            placeholder={placeholder}
+            className={baseStx}
+            disabled={disabled}
+            isOpen={isOpen}
+            onToggle={onToggle}
+            showLabel={showLabel}
+          />
         )}
       />
       <Popover.Portal
         onClickOutside={({ closePortal }) => closePortal()}
         render={({ setRef, togglePortal, isOpen, floatingStyles }) =>
           isOpen && (
-            <ul
+            <SelectList
               ref={setRef}
-              className={classNames(styles.selectList, listStx)}
+              className={baseStx}
               style={floatingStyles}
-            >
-              {options.map((option) => {
-                const handleClick = () => {
-                  togglePortal();
+              options={options}
+              width={mainWidth}
+              onItemClicked={(v) => {
+                togglePortal();
 
-                  if (!disabled) {
-                    setInternalValue(option.value);
-                    onChange?.(option.value);
-                  }
-                };
-                return (
-                  <li key={option.value} className={styles.selectItem}>
-                    <button
-                      type="button"
-                      className={styles.selectItemButton}
-                      onClick={handleClick}
-                      onKeyDown={(e) => {
-                        if (e.key === "Enter" || e.key === " ") {
-                          e.preventDefault();
-                          handleClick();
-                        }
-                      }}
-                      tabIndex={0}
-                    >
-                      {option.label}
-                    </button>
-                  </li>
-                );
-              })}
-            </ul>
+                if (!disabled) {
+                  setInternalValue(v);
+                  onChange?.(v);
+                }
+              }}
+            />
           )
         }
       />